refactor: extract assembleBroadcastEmail into shared utility
Moves assembleBroadcastEmail and buildPostCardsHtml out of send.ts into
src/resend/assembleBroadcastEmail.ts so the preview route (Phase 4) can
reuse them. Adds a preview option that replaces the Resend unsubscribe
placeholder with a visible label for in-browser preview rendering.
